chore: move pallet types to primitives
